# packages/api-client/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeMapAI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface FetchRepoRequest {
  /** GitHub repository URL */
  url: string;
  /** Branch to fetch (defaults to main/master) */
  branch?: string;
}

export type CodeNodeType = (typeof CodeNodeType)[keyof typeof CodeNodeType];

export const CodeNodeType = {
  root: "root",
  folder: "folder",
  file: "file",
  function: "function",
  class: "class",
  export: "export",
} as const;

export type CodeNodeComplexity =
  (typeof CodeNodeComplexity)[keyof typeof CodeNodeComplexity];

export const CodeNodeComplexity = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CodeNode {
  id: string;
  name: string;
  type: CodeNodeType;
  path: string;
  children?: CodeNode[];
  language?: string;
  size?: number;
  complexity?: CodeNodeComplexity;
  isEntryPoint?: boolean;
  content?: string;
}

export interface RepoStructure {
  repoName: string;
  rootNode: CodeNode;
  fileCount: number;
  languages: string[];
  entryPoints?: string[];
  totalSize?: number;
  sourceUrl?: string;
  branch?: string;
}

export interface ExplainNodeRequest {
  nodeId: string;
  nodeName: string;
  nodePath: string;
  nodeType: string;
  repoName: string;
  /** Brief description of the full repo structure for context */
  repoContext?: string;
  /** Optional file content to explain */
  content?: string;
  sourceUrl?: string;
  branch?: string;
}

export type ExplainNodeResponseComplexity =
  (typeof ExplainNodeResponseComplexity)[keyof typeof ExplainNodeResponseComplexity];

export const ExplainNodeResponseComplexity = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface ExplainNodeResponse {
  nodeId: string;
  /** What this file/function does */
  what: string;
  /** Why it exists */
  why: string;
  /** How it connects to other parts */
  connections: string;
  complexity?: ExplainNodeResponseComplexity;
  tips?: string[];
}

export interface ExplainFlowRequest {
  nodeIds: string[];
  nodePaths: string[];
  repoName: string;
  repoContext?: string;
  nodeContents?: (string | undefined)[];
  sourceUrl?: string;
  branch?: string;
}

export interface FlowStep {
  stepNumber: number;
  nodeId?: string;
  nodePath?: string;
  description: string;
  dataIn?: string;
  dataOut?: string;
}

export interface ExplainFlowResponse {
  steps: FlowStep[];
  summary: string;
}

export interface LearningPathRequest {
  repoName: string;
  rootNode: CodeNode;
  languages?: string[];
}

export interface LearningStep {
  order: number;
  nodeId?: string;
  nodePath?: string;
  reason: string;
  concept?: string;
}

export interface LearningPathResponse {
  steps: LearningStep[];
  summary: string;
  estimatedTime?: string;
}

export interface AskCodebaseRequest {
  question: string;
  repoName: string;
  repoContext?: string;
}

export type AskCodebaseResponseConfidence =
  (typeof AskCodebaseResponseConfidence)[keyof typeof AskCodebaseResponseConfidence];

export const AskCodebaseResponseConfidence = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface AskCodebaseResponse {
  answer: string;
  referencedNodeIds: string[];
  referencedPaths: string[];
  confidence?: AskCodebaseResponseConfidence;
}

export type UploadZipBody = {
  file: Blob;
};