# packages/api-contracts/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeMapAI API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Fetches and parses the structure of a GitHub repository or processes uploaded ZIP
 * @summary Fetch repository structure
 */
export const FetchRepoBody = zod.object({
  url: zod.string().describe("GitHub repository URL"),
  branch: zod
    .string()
    .optional()
    .describe("Branch to fetch (defaults to main\/master)"),
});

export const FetchRepoResponse = zod.object({
  repoName: zod.string(),
  rootNode: zod.object({
    id: zod.string(),
    name: zod.string(),
    type: zod.enum(["root", "folder", "file", "function", "class", "export"]),
    path: zod.string(),
    children: zod.array(zod.unknown()).optional(),
    language: zod.string().optional(),
    size: zod.number().optional(),
    complexity: zod.enum(["low", "medium", "high"]).optional(),
    isEntryPoint: zod.boolean().optional(),
    content: zod.string().optional(),
  }),
  fileCount: zod.number(),
  languages: zod.array(zod.string()),
  entryPoints: zod.array(zod.string()).optional(),
  totalSize: zod.number().optional(),
  sourceUrl: zod.string().optional(),
  branch: zod.string().optional(),
});

/**
 * Upload a ZIP file containing a codebase
 * @summary Upload ZIP file
 */
export const UploadZipBody = zod.object({
  file: zod.instanceof(File),
});

export const UploadZipResponse = zod.object({
  repoName: zod.string(),
  rootNode: zod.object({
    id: zod.string(),
    name: zod.string(),
    type: zod.enum(["root", "folder", "file", "function", "class", "export"]),
    path: zod.string(),
    children: zod.array(zod.unknown()).optional(),
    language: zod.string().optional(),
    size: zod.number().optional(),
    complexity: zod.enum(["low", "medium", "high"]).optional(),
    isEntryPoint: zod.boolean().optional(),
    content: zod.string().optional(),
  }),
  fileCount: zod.number(),
  languages: zod.array(zod.string()),
  entryPoints: zod.array(zod.string()).optional(),
  totalSize: zod.number().optional(),
  sourceUrl: zod.string().optional(),
  branch: zod.string().optional(),
});

/**
 * Get an AI explanation for a file or code element
 * @summary Explain a code node
 */
export const ExplainNodeBody = zod.object({
  nodeId: zod.string(),
  nodeName: zod.string(),
  nodePath: zod.string(),
  nodeType: zod.string(),
  repoName: zod.string(),
  repoContext: zod
    .string()
    .optional()
    .describe("Brief description of the full repo structure for context"),
  content: zod.string().optional().describe("Optional file content to explain"),
  sourceUrl: zod.string().url().optional(),
  branch: zod.string().optional(),
});

export const ExplainNodeResponse = zod.object({
  nodeId: zod.string(),
  what: zod.string().describe("What this file\/function does"),
  why: zod.string().describe("Why it exists"),
  connections: zod.string().describe("How it connects to other parts"),
  complexity: zod.enum(["low", "medium", "high"]).optional(),
  tips: zod.array(zod.string()).optional(),
});

/**
 * Generate step-by-step explanation of data flow between selected nodes
 * @summary Explain data flow between nodes
 */
export const ExplainFlowBody = zod.object({
  nodeIds: zod.array(zod.string()),
  nodePaths: zod.array(zod.string()),
  repoName: zod.string(),
  repoContext: zod.string().optional(),
  nodeContents: zod.array(zod.string().optional()).optional(),
  sourceUrl: zod.string().url().optional(),
  branch: zod.string().optional(),
});

export const ExplainFlowResponse = zod.object({
  steps: zod.array(
    zod.object({
      stepNumber: zod.number(),
      nodeId: zod.string().optional(),
      nodePath: zod.string().optional(),
      description: zod.string(),
      dataIn: zod.string().optional(),
      dataOut: zod.string().optional(),
    }),
  ),
  summary: zod.string(),
});

/**
 * Get AI-suggested order to understand the codebase
 * @summary Get suggested learning path
 */
export const GetLearningPathBody = zod.object({
  repoName: zod.string(),
  rootNode: zod.object({
    id: zod.string(),
    name: zod.string(),
    type: zod.enum(["root", "folder", "file", "function", "class", "export"]),
    path: zod.string(),
    children: zod.array(zod.unknown()).optional(),
    language: zod.string().optional(),
    size: zod.number().optional(),
    complexity: zod.enum(["low", "medium", "high"]).optional(),
    isEntryPoint: zod.boolean().optional(),
  }),
  languages: zod.array(zod.string()).optional(),
});

export const GetLearningPathResponse = zod.object({
  steps: zod.array(
    zod.object({
      order: zod.number(),
      nodeId: zod.string().optional(),
      nodePath: zod.string().optional(),
      reason: zod.string(),
      concept: zod.string().optional(),
    }),
  ),
  summary: zod.string(),
  estimatedTime: zod.string().optional(),
});

/**
 * Ask any question about the codebase and get AI answers with node references
 * @summary Ask a question about the codebase
 */
export const AskCodebaseBody = zod.object({
  question: zod.string(),
  repoName: zod.string(),
  repoContext: zod.string().optional(),
});

export const AskCodebaseResponse = zod.object({
  answer: zod.string(),
  referencedNodeIds: zod.array(zod.string()),
  referencedPaths: zod.array(zod.string()),
  confidence: zod.enum(["high", "medium", "low"]).optional(),
});