# packages/api-contracts/src/generated/types/codeNode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeMapAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CodeNodeComplexity } from "./codeNodeComplexity";
import type { CodeNodeType } from "./codeNodeType";

export interface CodeNode {
  id: string;
  name: string;
  type: CodeNodeType;
  path: string;
  children?: CodeNode[];
  language?: string;
  size?: number;
  complexity?: CodeNodeComplexity;
  isEntryPoint?: boolean;
  content?: string;
}