# packages/api-contracts/src/generated/types/explainFlowRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeMapAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExplainFlowRequest {
  nodeIds: string[];
  nodePaths: string[];
  repoName: string;
  repoContext?: string;
  nodeContents?: (string | undefined)[];
  sourceUrl?: string;
  branch?: string;
}