# packages/api-contracts/src/generated/types/explainNodeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeMapAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExplainNodeRequest {
  nodeId: string;
  nodeName: string;
  nodePath: string;
  nodeType: string;
  repoName: string;
  /** Brief description of the full repo structure for context */
  repoContext?: string;
  /** Optional file content to explain */
  content?: string;
  sourceUrl?: string;
  branch?: string;
}