# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CodeMapAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/explain-node:
    post:
      operationId: explainNode
      tags: [ai]
      summary: Explain a code node
      description: Get an AI explanation for a file or code element
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExplainNodeRequest"
      responses:
        "200":
          description: AI explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExplainNodeResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ai/explain-flow:
    post:
      operationId: explainFlow
      tags: [ai]
      summary: Explain data flow between nodes
      description: Generate step-by-step explanation of data flow between selected nodes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExplainFlowRequest"
      responses:
        "200":
          description: Flow explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExplainFlowResponse"
  /ai/learning-path:
    post:
      operationId: getLearningPath
      tags: [ai]
      summary: Get suggested learning path
      description: Get AI-suggested order to understand the codebase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LearningPathRequest"
      responses:
        "200":
          description: Learning path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningPathResponse"
  /ai/ask:
    post:
      operationId: askCodebase
      tags: [ai]
      summary: Ask a question about the codebase
      description: Ask any question about the codebase and get AI answers with node references
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskCodebaseRequest"
      responses:
        "200":
          description: Answer with node references
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AskCodebaseResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    CodeNode:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [root, folder, file, function, class, export]
        path:
          type: string
        children:
          type: array
          items:
            $ref: "#/components/schemas/CodeNode"
        language:
          type: string
        size:
          type: integer
        complexity:
          type: string
          enum: [low, medium, high]
        isEntryPoint:
          type: boolean
        content:
          type: string
          description: Optional source content for small text files
      required:
        - id
        - name
        - type
        - path
...
    ExplainNodeRequest:
      type: object
      properties:
        nodeId:
          type: string
        nodeName:
          type: string
        nodePath:
          type: string
        nodeType:
          type: string
        repoName:
          type: string
        repoContext:
          type: string
          description: Brief description of the full repo structure for context
        content:
          type: string
          description: Optional file content to explain
        sourceUrl:
          type: string
          format: uri
        branch:
          type: string
      required:
        - nodeId
        - nodeName
        - nodePath
        - nodeType
        - repoName
    ExplainNodeResponse:
      type: object
      properties:
        nodeId:
          type: string
        what:
          type: string
          description: What this file/function does
        why:
          type: string
          description: Why it exists
        connections:
          type: string
          description: How it connects to other parts
        complexity:
          type: string
          enum: [low, medium, high]
        tips:
          type: array
          items:
            type: string
      required:
        - nodeId
        - what
        - why
        - connections
    ExplainFlowRequest:
      type: object
      properties:
        nodeIds:
          type: array
          items:
            type: string
        nodePaths:
          type: array
          items:
            type: string
        repoName:
          type: string
        repoContext:
          type: string
        nodeContents:
          type: array
          items:
            type: string
            nullable: true
        sourceUrl:
          type: string
          format: uri
        branch:
          type: string
      required:
        - nodeIds
        - nodePaths
        - repoName
    ExplainFlowResponse:
      type: object
      properties:
        steps:
          type: array
          items:
            $ref: "#/components/schemas/FlowStep"
        summary:
          type: string
      required:
        - steps
        - summary
    FlowStep:
      type: object
      properties:
        stepNumber:
          type: integer
        nodeId:
          type: string
        nodePath:
          type: string
        description:
          type: string
        dataIn:
          type: string
        dataOut:
          type: string
      required:
        - stepNumber
        - description
    LearningPathRequest:
      type: object
      properties:
        repoName:
          type: string
        rootNode:
          $ref: "#/components/schemas/CodeNode"
        languages:
          type: array
          items:
            type: string
      required:
        - repoName
        - rootNode
    LearningPathResponse:
      type: object
      properties:
        steps:
          type: array
          items:
            $ref: "#/components/schemas/LearningStep"
        summary:
          type: string
        estimatedTime:
          type: string
      required:
        - steps
        - summary
    LearningStep:
      type: object
      properties:
        order:
          type: integer
        nodeId:
          type: string
        nodePath:
          type: string
        reason:
          type: string
        concept:
          type: string
      required:
        - order
        - reason
    AskCodebaseRequest:
      type: object
      properties:
        question:
          type: string
        repoName:
          type: string
        repoContext:
          type: string
      required:
        - question
        - repoName
    AskCodebaseResponse:
      type: object
      properties:
        answer:
          type: string
        referencedNodeIds:
          type: array
          items:
            type: string
        referencedPaths:
          type: array
          items:
            type: string
        confidence:
          type: string
          enum: [high, medium, low]
      required:
        - answer
        - referencedNodeIds
        - referencedPaths